some fixes
